feat(end-to-end): add visual regression tests for domain components and remove outdated tests
